Cleaner

Simple tools for common data cleaning problems.

Cleaner.CleanTable

Cleaner.generate_polished_names

Cleaner.polish_names

Cleaner.polish_names!

Cleaner.CleanTable — Type

CleanTable <: Tables.AbstractColumns

A Tables.jl implementation that stores column names and columns for Cleaner.jl internal use.

The default behavior of this type is to try to copy the columns of the original Tables implementation a.k.a: the source, but the user can call the second constructor specifiying copycols=false to override this behavior and try to use the original columns directly.

If the source column type is not mutable, this will end up in errors.

Constructors

CleanTable(names::Vector{Symbol}, cols; copycols::Bool=true)
CleanTable(table; copycols::Bool=true)
CleanTable(table::CleanTable; copycols::Bool=true)

Cleaner.generate_polished_names — Method

generate_polished_names(names; style::Symbol=:snake_case)

Return a vector of symbols containing new names that are unique and formated using the style selected.

Cleaner.polish_names! — Method

polish_names!(table::CleanTable; style::Symbol=:snake_case)

Return a CleanTable where column names have been replaced to be unique and formated using the style selected.

Styles

snake_case

camelCase

Cleaner.polish_names — Method

polish_names(table; style=:snake_case)

Return a CleanTable with copied columns having column names replaced to be unique and formated using the style selected.

Styles

snake_case

camelCase