Functions

Multi-threading support

The following functions will try to use multiple threads if possible when there are at least 2 columns and 1 million rows:

CleanTable constructor when copycols=true

All compact functions

delete_const_columns and delete_const_columns!

reinfer_schema and reinfer_schema!

Index

Cleaner.compact_columns

Cleaner.compact_columns!

Cleaner.compact_rows

Cleaner.compact_rows!

Cleaner.compact_table

Cleaner.compact_table!

Cleaner.delete_const_columns

Cleaner.delete_const_columns!

Cleaner.generate_polished_names

Cleaner.polish_names

Cleaner.polish_names!

Cleaner.reinfer_schema

Cleaner.reinfer_schema!

Cleaner.rename

Cleaner.rename!

Cleaner.row_as_names

Cleaner.row_as_names!

Cleaner.size

Summarize information

Cleaner.size — Function

size(table::CleanTable)

Returns a tuple containing the number of rows and columns of the given table.

Working with column names

Cleaner.rename — Function

rename(ct::CleanTable, names::Vector{Symbol})

Creates a CleanTable with copied columns and changes its column names to be the names passed as argument.

Cleaner.rename! — Function

rename!(ct::CleanTable, names::Vector{Symbol})

Changes in-place the column names of a CleanTable to be the names passed as argument.

Cleaner.generate_polished_names — Function

generate_polished_names(names; style::Symbol=:snake_case)

Return a vector of symbols containing new names that are unique and formated using the style selected.

Cleaner.polish_names — Function

polish_names(table; style=:snake_case)

Create and return a CleanTable with copied columns having column names replaced to be unique and formated using the style selected.

Styles

snake_case

camelCase

Cleaner.polish_names! — Function

polish_names!(table::CleanTable; style::Symbol=:snake_case)

Return a CleanTable where column names have been replaced to be unique and formated using the style selected.

Styles

snake_case

camelCase

Cleaner.row_as_names — Function

row_as_names(table, i::Int; remove::Bool=false)

Creates a CleanTable with copied columns and renames the table using row i as new names and removes in-place all the rows above row i if remove=true.

Default behavior is to remove rows above row i.

Cleaner.row_as_names! — Function

row_as_names!(table::CleanTable, i::Int; remove::Bool=true)

Renames the table using row i as new names and removes in-place all the rows above row i if remove=true.

Default behavior is to remove rows above row i.

Row/Column removal

Cleaner.compact_columns — Function

compact_columns(table; empty_values::Vector=[])

Creates a CleanTable with copied columns and removes from it all columns filled entirely by missing and empty_values.

Cleaner.compact_columns! — Function

compact_columns!(table::CleanTable; empty_values::Vector=[])

Removes in-place from a CleanTable all columns filled entirely by missing and empty_values.

Cleaner.compact_rows — Function

compact_rows(table; empty_values::Vector=[])

Creates a CleanTable with copied columns and removes from it all rows filled entirely by missing and empty_values.

Cleaner.compact_rows! — Function

compact_rows!(table::CleanTable; empty_values::Vector=[])

Removes in-place from a CleanTable all rows filled entirely by missing and empty_values.

Cleaner.compact_table — Function

compact_table(table; empty_values::Vector=[])

Creates a CleanTable with copied columns and removes from it all rows and columns filled entirely by missing and empty_values.

Cleaner.compact_table! — Function

compact_table!(table::CleanTable; empty_values::Vector=[])

Removes in-place from a CleanTable all rows and columns filled entirely by missing and empty_values.

Cleaner.delete_const_columns — Function

delete_const_columns(table)

Creates a CleanTable with copied columns and removes each column filled with just a constant value.

Cleaner.delete_const_columns! — Function

delete_const_columns!(table::CleanTable)

Removes in-place from a CleanTable each column filled with just a constant value.

Modifiying table schema

Cleaner.reinfer_schema — Function

reinfer_schema(table; max_types::Int=3)

Creates a CleanTable with copied columns and tries to minimize the amount of element types for each column without making the column type Any.

For this, will try to make the column of type Union with up to maxtypes and internally use ```Base.promotetypejoin``` on all numeric types. If not possible, leaves the column as-is.

Cleaner.reinfer_schema! — Function

reinfer_schema!(table::CleanTable; max_types::Int=3)

Tries to minimize the amount of element types for each column without making the column type Any.

For this, will try to make the column of type Union with up to maxtypes and internally use ```Base.promotetypejoin``` on all numeric types. If not possible, leaves the column as-is.